+++ packages/babel-runtime-corejs2/helpers/esm/temporalRef.js
@@ -1,1 +1,5 @@
-export { default } from "../temporalRef/_index.mjs"
+import undef from "./temporalUndefined.js";
+import err from "./tdz.js";
+export default function _temporalRef(val, name) {
+  return val === undef ? err(name) : val;
+}
