+++ packages/babel-runtime/helpers/esm/iterableToArray.js
@@ -1,1 +1,3 @@
-export { default } from "../iterableToArray/_index.mjs"
+export default function _iterableToArray(iter) {
+  if (typeof Symbol !== "undefined" && Symbol.iterator in Object(iter)) return Array.from(iter);
+}
